Better handling of unicode supplementary chars
Supplementary characters require 2 java/ kotlin
chars to store but are rendered as a single
character. This caused problems for a lot of
places that assumed length = number of chars.

Replaced the buffer in TextSurface to be an array
of codepoints (int) rather than a plain String (which
would have an undefined length if surrogate pairs are
used). Iterate through the codepoints rather than chars.

Note this is more consistent but not necessarily
always better than to measuring these as 2 chars.
Text measurement is complicated and I'll add more
context to #4