Update dependencies to fix build
Using the previous dependency versions, the build always failed during
the documentation tasks; this was due to an incompatibility between
Kotlin and Gradle (see https://github.com/Kotlin/dokka/issues/428).

Update Kotlin to 1.4.30 (not the latest version, but the latest version
available in IDEA). KTLint could not be updated to 0.41.0 because of
some weird behavior with the plugin (sample/Main.kt indentation is
marked as incorrect).

Update the configuration blocks, as well (the `dokka` block is no more).